requirements.txt
certifi==2021.5.30
cffi @ file:///tmp/build/80754af9/cffi_1625814693446/work
cycler==0.10.0
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1612282414123/work
matplotlib @ file:///tmp/build/80754af9/matplotlib-suite_1613407855456/work
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603479632437/work
olefile==0.46
Pillow @ file:///tmp/build/80754af9/pillow_1625655818400/work
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
six @ file:///tmp/build/80754af9/six_1623709665295/work
torch==1.9.0
torchaudio==0.9.0
torchvision==0.2.1
tornado @ file:///tmp/build/80754af9/tornado_1606942283357/work
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1624965014186/work
